Factor MockUIPush out of OnlineLeaguePlayerLanding.test
